Lê só o byte de retorno das funções da EasyInner.dll
A DLL preenche apenas o byte baixo do retorno; declarado como int, um
comando OK chegava como 196877312 (0x0BBC1C00) e a ponte abortava ao
abrir a porta 3570 achando que tinha falhado.

+++ catraca-ponte/EasyInner.cs
@@ -9,30 +9,34 @@ namespace CatracaPonte;
 /// seguem o EasyInner.cs dos exemplos em C# da Topdata; se a sua versão da
 /// DLL tiver alguma assinatura diferente, ajuste aqui (é o único lugar que
 /// fala com a DLL).
+///
+/// Os retornos são declarados como byte: a DLL só preenche o byte baixo do
+/// registrador de retorno, e ler como int trazia lixo nos bytes de cima (ex:
+/// 196877312 = 0x0BBC1C00 para um "comando OK").
 /// </summary>
 internal static class EasyInner
 {
     private const string Dll = "EasyInner.dll";
 
-    public const int RetComandoOk = 0;
+    public const byte RetComandoOk = 0;
 
     [DllImport(Dll, CallingConvention = CallingConvention.StdCall)]
-    public static extern int DefinirTipoConexao(byte tipo);
+    public static extern byte DefinirTipoConexao(byte tipo);
 
     [DllImport(Dll, CallingConvention = CallingConvention.StdCall)]
-    public static extern int AbrirPortaComunicacao(int porta);
+    public static extern byte AbrirPortaComunicacao(int porta);
 
     [DllImport(Dll, CallingConvention = CallingConvention.StdCall)]
     public static extern void FecharPortaComunicacao();
 
     [DllImport(Dll, CallingConvention = CallingConvention.StdCall)]
-    public static extern int Ping(int inner);
+    public static extern byte Ping(int inner);
 
     [DllImport(Dll, CallingConvention = CallingConvention.StdCall)]
-    public static extern int PingOnLine(int inner);
+    public static extern byte PingOnLine(int inner);
 
     [DllImport(Dll, CallingConvention = CallingConvention.StdCall)]
-    public static extern int DefinirPadraoCartao(byte padrao);
+    public static extern byte DefinirPadraoCartao(byte padrao);
 
     [DllImport(Dll, CallingConvention = CallingConvention.StdCall)]
     public static extern void ConfigurarInnerOnLine();
@@ -59,35 +63,35 @@ internal static class EasyInner
     public static extern void HabilitarTeclado(byte habilitar, byte ecoar);
 
     [DllImport(Dll, CallingConvention = CallingConvention.StdCall)]
-    public static extern int EnviarConfiguracoes(int inner);
+    public static extern byte EnviarConfiguracoes(int inner);
 
     [DllImport(Dll, CallingConvention = CallingConvention.StdCall, CharSet = CharSet.Ansi)]
-    public static extern int EnviarMensagemPadraoOnLine(int inner, int exibirData, string mensagem);
+    public static extern byte EnviarMensagemPadraoOnLine(int inner, int exibirData, string mensagem);
 
     [DllImport(Dll, CallingConvention = CallingConvention.StdCall)]
-    public static extern int EnviarFormasEntradasOnLine(
+    public static extern byte EnviarFormasEntradasOnLine(
         int inner, byte qtdeDigitosTeclado, byte ecoDisplay, byte formaEntrada, byte tempoTeclado, byte posicaoCursorTeclado);
 
     [DllImport(Dll, CallingConvention = CallingConvention.StdCall, CharSet = CharSet.Ansi)]
-    public static extern int ReceberDadosOnLine(
+    public static extern byte ReceberDadosOnLine(
         int inner, ref byte origem, ref byte complemento, StringBuilder cartao,
         ref byte dia, ref byte mes, ref byte ano, ref byte hora, ref byte minuto, ref byte segundo);
 
     [DllImport(Dll, CallingConvention = CallingConvention.StdCall)]
-    public static extern int LiberarCatracaEntrada(int inner);
+    public static extern byte LiberarCatracaEntrada(int inner);
 
     [DllImport(Dll, CallingConvention = CallingConvention.StdCall)]
-    public static extern int LiberarCatracaEntradaInvertida(int inner);
+    public static extern byte LiberarCatracaEntradaInvertida(int inner);
 
     [DllImport(Dll, CallingConvention = CallingConvention.StdCall)]
-    public static extern int LiberarCatracaSaida(int inner);
+    public static extern byte LiberarCatracaSaida(int inner);
 
     [DllImport(Dll, CallingConvention = CallingConvention.StdCall)]
-    public static extern int LiberarCatracaSaidaInvertida(int inner);
+    public static extern byte LiberarCatracaSaidaInvertida(int inner);
 
     [DllImport(Dll, CallingConvention = CallingConvention.StdCall)]
-    public static extern int LiberarCatracaDoisSentidos(int inner);
+    public static extern byte LiberarCatracaDoisSentidos(int inner);
 
     [DllImport(Dll, CallingConvention = CallingConvention.StdCall)]
-    public static extern int AcionarBipLongo(int inner);
+    public static extern byte AcionarBipLongo(int inner);
 }
